chore(scroll-anchor): production cleanup — move hook, repurpose scrollDebug as dev tool
Cleanup pass before PR. No behavior change for the parts that already
worked; threads remain regressed (documented as Session 22 follow-up,
to be fixed in a fresh session).

File moves:
  src/components/message/useScrollAnchor.ts → src/hooks/ui/useScrollAnchor.ts
    (per project convention; lives next to useScrollTracking)
  src/components/message/__scrollDebug.ts   → src/dev/scrollDebug.ts
    (kept as a permanent dev tool, no longer "TEMPORARY DEBUG"; header
     rewritten to document how to wire it in for ad-hoc debugging)

Production code stripped of debug calls:
  MessageList.tsx — removed scroll-recorder attach effect, log in
    handleBottomStateChange, log in handleRangeChanged
  MessageService.ts — removed scrollDebug import + addMessage log
  useScrollAnchor.ts — removed all scrollDebug.log calls + commented-out
    absorb branch + now-unused BACKWARD_JUMP_THRESHOLD_PX constant
  DirectMessage.tsx — already clean from earlier commit

Docs:
  bugs/2026-05-24-... — Session 22 added documenting the cleanup and
    the thread regression discovered during pre-PR sanity check. Status
    reverted from fixed-pending-cleanup → in-progress because threads
    need a fix before PR. Current-behavior table updated with the
    thread regression rows. Full pickup plan for next session included.
  docs/features/messages/scroll-anchoring.md — file paths updated to
    reflect new locations; regression-debug instructions updated to
    point at src/dev/scrollDebug.ts.

Type-check clean.

Next session: extend useScrollAnchor with optional queryKeyPrefix
param so ThreadPanel can subscribe to thread-messages query; add
explicit messageListRef.current?.scrollToBottom() to ThreadPanel's
submit handler. Then test threads end-to-end, then PR.
